requirements.txt
protobuf
Twisted
pyOpenSSL
pyocb
service_identity
requests
